Reworks tests for client arf-to-json
